Lektion3 another commit

+++ Lektioner/Lektion3.py
@@ -1,20 +1,29 @@
 name = input("What is your name?")
 age = int(input("Hur gammal är du?"))
+#int cause whole number
 height = float(input("Hur lång är du i cm?"))
+#float cause decimals possible
 
 print("Hej", name,"du är",age,"år gammal och",height,"cm lång")
+#prints name, age, height
 
 ordtal=input("Input number")
 print(5*ordtal)
+#prints inputinputinputinputinput
 
 
 tal= int(input ("Input number"))
 print(5 * tal)
+#since its an int this time and not a string it will actually print 5*inputted number
 
 fee= 358.3
+#Specify price per hour
 namn=input("Hej! Välkommen till BananaFitness, vad är ditt namn?")
-hours =int(input("Hur många timmar i veckan tänkte du att träna med oss?"))
+#grabs name
+hours=float(input("Hur många timmar i veckan tänkte du att träna med oss?"))
+#grabs hours per week, float in case half hours
 print("Vad kul", namn+"!", "För att träna", hours, "timmar i veckan så skulle det kosta", hours*fee,"kr per vecka")
+#prints some nice words while saying their name, amount of hours and cost per week by doing hours * fee
 print (round (1337.1337,2))
 print (round (1337.1337))
 #Rounds up numbers, if you type (1337.1337) and round it it will round down, however if you add a comma with a number it will
